完善了通过支付宝扫码支付的功能。

+++ src/main/java/com/lei/mall/model/entity/PurchaseRecord.java
@@ -3,6 +3,7 @@ package com.lei.mall.model.entity;
 import com.baomidou.mybatisplus.annotation.*;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import lombok.Data;
 
@@ -42,7 +43,7 @@ public class PurchaseRecord implements Serializable {
      * 支付金额
      */
     @TableField(value = "paymentAmount")
-    private Integer paymentAmount;
+    private BigDecimal paymentAmount;
 
     /**
      * 支付时间
